refactor: rename mcpBootstrap module to mcpInvites
The surviving files (send_deployment_invites, configure_defaults, the
invite landing page, the enrollment matcher, metrics) are about agent-
driven device deployment, not tenant bootstrap. Rename the module to
match what it actually is.

Tool descriptions also de-bootstrapped — references to deleted tools
(create_tenant / verify_tenant / attach_payment_method) and the
PAYMENT_REQUIRED gate replaced with active-tenant phrasing.

Refs: Phase 6 of docs/superpowers/plans/2026-04-29-mcp-bootstrap-cleanup.md
