+++ apps/api/src/modules/mcpInvites/tools/configureDefaults.test.ts
