+++ apps/api/src/routes/agents/enrollment.ts
@@ -21,7 +21,7 @@ import { enrollSchema } from './schemas';
 import { generateAgentId, generateApiKey, issueMtlsCertForDevice } from './helpers';
 import { queueWarrantySyncForDevice } from '../../services/warrantyWorker';
 import { dispatchHook } from '../../services/partnerHooks';
-import { matchDeploymentInviteOnEnrollment } from '../../modules/mcpBootstrap/matchInviteOnEnrollment';
+import { matchDeploymentInviteOnEnrollment } from '../../modules/mcpInvites/matchInviteOnEnrollment';
 
 export const enrollmentRoutes = new Hono();
 const ENROLLMENT_RATE_LIMIT = 10;
